+++ src/bp/core/bots/bot-monitoring-service.ts
@@ -1,8 +1,7 @@
 import 'bluebird-global'
 import { Logger, LoggerEntry } from 'botpress/sdk'
 import { ConfigProvider } from 'core/config'
-import { JobService } from 'core/distributed'
-import { makeRedisKey } from 'core/distributed/redis'
+import { JobService, makeRedisKey } from 'core/distributed'
 import { LogsRepository } from 'core/logger'
 import { TYPES } from 'core/types'
 import { inject, injectable, tagged } from 'inversify'
