+++ src/bp/core/bots/bot-service.ts
@@ -7,8 +7,7 @@ import { TYPES } from 'core/app/types'
 import { FileContent, GhostService, ReplaceContent } from 'core/bpfs'
 import { CMSService } from 'core/cms'
 import { ConfigProvider } from 'core/config'
-import { JobService } from 'core/distributed'
-import { makeRedisKey } from 'core/distributed/redis'
+import { JobService, makeRedisKey } from 'core/distributed'
 import { PersistedConsoleLogger } from 'core/logger'
 import { MigrationService } from 'core/migration'
 import { extractArchive } from 'core/misc/archive'
