+++ src/bp/core/dialog/state-manager.ts
@@ -2,8 +2,7 @@ import * as sdk from 'botpress/sdk'
 import { TYPES } from 'core/app/types'
 import { BotpressConfig, ConfigProvider } from 'core/config'
 import Database from 'core/database'
-import { JobService } from 'core/distributed'
-import { makeRedisKey } from 'core/distributed/redis'
+import { JobService, makeRedisKey } from 'core/distributed'
 import { KeyValueStore } from 'core/kvs'
 import { ChannelUserRepository } from 'core/users'
 import { inject, injectable, tagged } from 'inversify'
