+++ src/bp/core/distributed/index.ts
@@ -1,1 +1,2 @@
 export * from './job-service'
+export * from './redis'
